+++ src/gui/DatabaseOpenWidget.cpp
@@ -72,11 +72,9 @@ DatabaseOpenWidget::DatabaseOpenWidget(QWidget* parent)
 
     connect(m_ui->buttonBrowseFile, SIGNAL(clicked()), SLOT(browseKeyFile()));
 
-    auto okBtn = m_ui->buttonBox->button(QDialogButtonBox::Ok);
-    okBtn->setText(tr("Unlock"));
-    okBtn->setDefault(true);
-    connect(m_ui->buttonBox, SIGNAL(accepted()), SLOT(openDatabase()));
-    connect(m_ui->buttonBox, SIGNAL(rejected()), SLOT(reject()));
+    m_ui->unlockButton->setDefault(true);
+    connect(m_ui->unlockButton, SIGNAL(clicked()), SLOT(openDatabase()));
+    connect(m_ui->closeButton, SIGNAL(clicked()), SLOT(reject()));
 
     m_ui->addKeyFileLinkLabel->setText(QStringLiteral("<a href=\"#\" style=\"text-decoration: underline\">%1</a>")
                                            .arg(tr("I have a key file").toHtmlEscaped()));
@@ -307,6 +305,7 @@ void DatabaseOpenWidget::load(const QString& filename)
     m_ui->enableQuickUnlockCheckBox->setChecked(isQuickUnlockAvailable()
                                                 && config()->get(Config::Security_QuickUnlock).toBool());
     m_ui->enableQuickUnlockCheckBox->setVisible(isQuickUnlockAvailable());
+    m_ui->enableQuickUnlockLabel->setVisible(isQuickUnlockAvailable());
 
     // Do initial auto-poll
     pollHardwareKey();
@@ -370,20 +369,16 @@ void DatabaseOpenWidget::openDatabase()
     if (ok) {
         // Warn user about minor version mismatch to halt loading if necessary
         if (m_db->hasMinorVersionMismatch()) {
-            QScopedPointer<QMessageBox> msgBox(new QMessageBox(this));
-            msgBox->setIcon(QMessageBox::Warning);
-            msgBox->setWindowTitle(tr("Database Version Mismatch"));
-            msgBox->setText(tr("The database you are trying to open was most likely\n"
-                               "created by a newer version of KeePassXC.\n\n"
-                               "You can try to open it anyway, but it may be incomplete\n"
-                               "and saving any changes may incur data loss.\n\n"
-                               "We recommend you update your KeePassXC installation."));
-            auto btn = msgBox->addButton(tr("Open database anyway"), QMessageBox::ButtonRole::AcceptRole);
-            msgBox->setDefaultButton(btn);
-            msgBox->addButton(QMessageBox::Cancel);
-            msgBox->layout()->setSizeConstraint(QLayout::SetMinimumSize);
-            msgBox->exec();
-            if (msgBox->clickedButton() != btn) {
+            const auto answer = MessageBox::warning(this,
+                                                    tr("Database Version Mismatch"),
+                                                    tr("The database you are trying to open was most likely\n"
+                                                       "created by a newer version of KeePassXC.\n\n"
+                                                       "You can try to open it anyway, but it may be incomplete\n"
+                                                       "and saving any changes may incur data loss.\n\n"
+                                                       "We recommend you update your KeePassXC installation."),
+                                                    MessageBox::OpenAnyway | MessageBox::Cancel,
+                                                    MessageBox::OpenAnyway);
+            if (answer != MessageBox::OpenAnyway) {
                 m_db.reset(new Database());
                 m_db->open(m_filename, nullptr, &error);
 
@@ -407,19 +402,17 @@ void DatabaseOpenWidget::openDatabase()
         clearForms();
     } else {
         if (!isOnQuickUnlockScreen() && m_ui->editPassword->text().isEmpty() && !m_retryUnlockWithEmptyPassword) {
-            QScopedPointer<QMessageBox> msgBox(new QMessageBox(this));
-            msgBox->setIcon(QMessageBox::Critical);
-            msgBox->setWindowTitle(tr("Unlock failed and no password given"));
-            msgBox->setText(tr("Unlocking the database failed and you did not enter a password.\n"
-                               "Do you want to retry with an \"empty\" password instead?\n\n"
-                               "To prevent this error from appearing, you must go to "
-                               "\"Database Settings / Security\" and reset your password."));
-            auto btn = msgBox->addButton(tr("Retry with empty password"), QMessageBox::ButtonRole::AcceptRole);
-            msgBox->setDefaultButton(btn);
-            msgBox->addButton(QMessageBox::Cancel);
-            msgBox->exec();
+            const auto answer =
+                MessageBox::critical(this,
+                                     tr("Unlock failed and no password given"),
+                                     tr("Unlocking the database failed and you did not enter a password.\n"
+                                        "Do you want to retry with an \"empty\" password instead?\n\n"
+                                        "To prevent this error from appearing, you must go to "
+                                        "\"Database Settings / Security\" and reset your password."),
+                                     MessageBox::RetryWithEmptyPassword | MessageBox::Cancel,
+                                     MessageBox::RetryWithEmptyPassword);
 
-            if (msgBox->clickedButton() == btn) {
+            if (answer == MessageBox::RetryWithEmptyPassword) {
                 m_retryUnlockWithEmptyPassword = true;
                 setUserInteractionLock(false);
                 openDatabase();
@@ -477,22 +470,18 @@ QSharedPointer<CompositeKey> DatabaseOpenWidget::buildDatabaseKey()
         }
         if (key->type() != FileKey::KeePass2XMLv2 && key->type() != FileKey::Hashed
             && !config()->get(Config::Messages_NoLegacyKeyFileWarning).toBool()) {
-            QMessageBox legacyWarning;
-            legacyWarning.setWindowTitle(tr("Old key file format"));
-            legacyWarning.setText(tr("You are using an old key file format which KeePassXC may<br>"
-                                     "stop supporting in the future.<br><br>"
-                                     "Please consider generating a new key file by going to:<br>"
-                                     "<strong>Database &gt; Database Security &gt; Change Key File.</strong><br>"));
-            legacyWarning.setIcon(QMessageBox::Icon::Warning);
-            legacyWarning.addButton(QMessageBox::Ok);
-            legacyWarning.setDefaultButton(QMessageBox::Ok);
-            legacyWarning.setCheckBox(new QCheckBox(tr("Don't show this warning again")));
-
-            connect(legacyWarning.checkBox(), &QCheckBox::stateChanged, this, [](int state) {
-                config()->set(Config::Messages_NoLegacyKeyFileWarning, state == Qt::CheckState::Checked);
-            });
-
-            legacyWarning.exec();
+            QScopedPointer<QCheckBox> dontAskAgain(new QCheckBox(tr("Don't show this warning again")));
+            MessageBox::warning(this,
+                                tr("Old key file format"),
+                                tr("You are using an old key file format which KeePassXC may<br>"
+                                   "stop supporting in the future.<br><br>"
+                                   "Please consider generating a new key file by going to:<br>"
+                                   "<strong>Database &gt; Database Security &gt; Change Key File.</strong><br>"),
+                                MessageBox::Ok,
+                                MessageBox::Ok,
+                                MessageBox::None,
+                                dontAskAgain.data());
+            config()->set(Config::Messages_NoLegacyKeyFileWarning, dontAskAgain->isChecked());
         }
         databaseKey->addKey(key);
         lastKeyFiles.insert(m_filename, keyFilename);